DDF-4249 Move UI security code into sub-packages to better separate concerns
