+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/forms/data/QueryTemplateMetacard.java
@@ -31,7 +31,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.codice.ddf.catalog.ui.security.AccessControlUtil;
+import org.codice.ddf.catalog.ui.security.accesscontrol.AccessControlUtil;
 
 /**
  * Metacard used for storing query templates in the catalog. Should not be used as a resource.
