+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/forms/security/SystemTemplatePolicyPlugin.java
@@ -24,7 +24,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Predicate;
-import org.codice.ddf.catalog.ui.security.AccessControlSecurityConfiguration;
+import org.codice.ddf.catalog.ui.security.accesscontrol.AccessControlSecurityConfiguration;
 
 public class SystemTemplatePolicyPlugin extends AbstractPolicyPlugin {
   private final AccessControlSecurityConfiguration config;
