+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/security/accesscontrol/AccessControlPolicyPlugin.java
@@ -11,9 +11,9 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.ui.security;
+package org.codice.ddf.catalog.ui.security.accesscontrol;
 
-import static org.codice.ddf.catalog.ui.security.AccessControlUtil.CONTAINS_ACL_ATTRIBUTES;
+import static org.codice.ddf.catalog.ui.security.accesscontrol.AccessControlUtil.CONTAINS_ACL_ATTRIBUTES;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Sets;
