+++ catalog/ui/catalog-ui-search/src/main/java/org/codice/ddf/catalog/ui/security/faceting/FacetWhitelistConfiguration.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.ui.security;
+package org.codice.ddf.catalog.ui.security.faceting;
 
 import java.util.Collections;
 import java.util.List;
