+++ catalog/ui/catalog-ui-search/src/test/java/org/codice/ddf/catalog/ui/config/ConfigurationApplicationTest.java
@@ -35,7 +35,7 @@ import java.util.Map;
 import java.util.ResourceBundle;
 import java.util.Set;
 import org.apache.commons.io.IOUtils;
-import org.codice.ddf.catalog.ui.security.FacetWhitelistConfiguration;
+import org.codice.ddf.catalog.ui.security.faceting.FacetWhitelistConfiguration;
 import org.codice.ddf.platform.util.uuidgenerator.UuidGenerator;
 import org.codice.proxy.http.HttpProxyService;
 import org.junit.Before;
