+++ catalog/ui/catalog-ui-search/src/test/java/org/codice/ddf/catalog/ui/security/accesscontrol/AclTestSupportTest.java
@@ -11,9 +11,9 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.ui.security;
+package org.codice.ddf.catalog.ui.security.accesscontrol;
 
-import static org.codice.ddf.catalog.ui.security.AclTestSupport.metacardFromAttributes;
+import static org.codice.ddf.catalog.ui.security.accesscontrol.AclTestSupport.metacardFromAttributes;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
