+++ catalog/ui/catalog-ui-search/src/test/java/org/codice/ddf/catalog/ui/security/faceting/FacetAttributePluginTest.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.catalog.ui.security;
+package org.codice.ddf.catalog.ui.security.faceting;
 
 import static ddf.catalog.Constants.EXPERIMENTAL_FACET_PROPERTIES_KEY;
 import static org.hamcrest.CoreMatchers.is;
